requirements.txt
aiohttp==3.7.4
websockets>=10.0.0,<11.0.0 
requests==2.25.1 
python-a2s>=1.3.0
discord.py==2.4.0
mysql-connector-python
aiomysql==0.0.22        # For asynchronous MySQL connections
asyncpg
nest_asyncio==1.5.6




